Refactor: Data모듈 import 오류 해결
