+++ data/build.gradle.kts
@@ -38,8 +38,13 @@ subprojects {
         implementation("com.google.dagger:hilt-android:2.48.1")
         kapt("com.google.dagger:hilt-compiler:2.48.1")
         implementation("androidx.core:core-ktx:1.13.1")
-
+        implementation("androidx.room:room-runtime:2.6.1")
+        kapt("androidx.room:room-compiler:2.6.1")
+        implementation("androidx.room:room-ktx:2.6.1")
         implementation(project(":common:util"))
+        implementation(project(":domain"))
+        implementation(project(":domain:group"))
+
     }
 
     android {
