+++ data/group/build.gradle.kts
@@ -1,3 +1,10 @@
 android {
     namespace = "com.kappzzang.jeongsan.group"
 }
+
+dependencies {
+    implementation(project(":data"))
+    implementation(project(":domain:group"))
+    implementation(project(":domain:expense"))
+    implementation(project(":data:data-expense"))
+}
