+++ data/group/src/main/java/com/kappzzang/jeongsan/datasource/group/GroupDao.kt
@@ -4,7 +4,7 @@ import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
-import com.kappzzang.jeongsan.data.entity.GroupEntity
+import com.kappzzang.jeongsan.entity.GroupEntity
 
 @Dao
 interface GroupDao {
@@ -14,12 +14,12 @@ interface GroupDao {
     @Delete
     fun deleteGroup(groupEntity: GroupEntity)
 
-    @Query("SELECT * FROM `${GroupContract.GroupEntity.TABLE_NAME}` WHERE is_completed = 0")
+    @Query("SELECT * FROM `${GroupContract.GroupEntity.TABLE_NAME}` WHERE ${GroupContract.GroupEntity.COLUMN_IS_COMPLETED} = 0")
     fun getProgressingGroup(): List<GroupEntity>
 
-    @Query("SELECT * FROM `${GroupContract.GroupEntity.TABLE_NAME}` WHERE is_completed = 1")
+    @Query("SELECT * FROM `${GroupContract.GroupEntity.TABLE_NAME}` WHERE ${GroupContract.GroupEntity.COLUMN_IS_COMPLETED} = 1")
     fun getDoneGroup(): List<GroupEntity>
 
-    @Query("SELECT * FROM `${GroupContract.GroupEntity.TABLE_NAME}` WHERE id = :groupId")
+    @Query("SELECT * FROM `${GroupContract.GroupEntity.TABLE_NAME}` WHERE ${GroupContract.GroupEntity.COLUMN_ID} = :groupId")
     fun inquireGroupInfo(groupId: Long): List<GroupEntity>
 }
