+++ data/group/src/main/java/com/kappzzang/jeongsan/datasource/member/MemberDatabase.kt
@@ -2,7 +2,7 @@ package com.kappzzang.jeongsan.datasource.member
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.kappzzang.jeongsan.data.entity.MemberEntity
+import com.kappzzang.jeongsan.entity.MemberEntity
 
 @Database(entities = [MemberEntity::class], version = 1)
 abstract class MemberDatabase : RoomDatabase() {
